Añadido caso de uso para añadir un cliente.
