+++ src/com/iesam/chispas/Productos.java
@@ -1,3 +1,5 @@
+package com.iesam.chispas;
+
 public class Productos {
     private String cod;
     private String nombre;
